+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Clipboard actions in contextmenu and visual mode now work in shadow roots and iframes
 - Download image pointer action now works in shadow roots and iframes
 - Action "clickOnSearch" not working if the page was scrolled
+- Follow mode and related actions not working on chrome:// pages due to querySelectorAll oddities
 
 ### Security